requirements.txt
streamlit
pandasai
pandasai[connectors]
pandasai[langchain]
openai
llama-cpp-python